Delay displaying until script it loaded
