+++ app/templates/bing-map.hbs
@@ -1,0 +1,5 @@
+{{#load-bing-maps}}
+  {{#if loaded}}
+    {{yield (hash Microsoft=microsoft) }}
+  {{/if}}
+{{/load-bing-maps}}
