+++ app/templates/components/bing-map.hbs
@@ -1,0 +1,8 @@
+{{#if isLoaded}}
+  {{private-bing-map
+    pins=pins
+    options=options
+    infoBoxes=infoBoxes
+    Microsoft=microsoft
+  }}
+{{/if}}
