+++ index.js
@@ -2,10 +2,5 @@
 'use strict';
 
 module.exports = {
-  name: 'ember-bing-maps',
-  contentFor(type) {
-    if (type === 'body') {
-      return `<script type="text/javascript" src="https://www.bing.com/api/maps/mapcontrol"></script>`;
-    }
-  }
+  name: 'ember-bing-maps'
 };
